server/pyproject.toml
[project]
name = "hobheap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "kernels>=0.9.0",
    "langgraph>=0.6.3",
    "pinecone>=7.3.0",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "sqlalchemy>=2.0.42",
    "transformers>=4.55.*",
    "torch==2.7.1",
    "accelerate>=1.9.0",
  "triton>=3.3.1; sys_platform != 'win32'",
    "openai>=1.99.1",
    "transformers>=4.55.0",
    "llama-cpp-python>=0.3.15",
  "alembic>=1.13.2",
  "pyjwt>=2.9.0",
  "aiosqlite>=0.20.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "streamlit>=1.47.1",
  "pytest>=8.3.3",
  "httpx>=0.27.2",
  "pytest-asyncio>=0.24.0",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
